burst should be set near max-rate
Burst value for TCP traffic should be set as 80% of desired bandwidth
limit value. For example, if the bandwidth limit is set to 1000kbps
then enough burst value will be 800kbit. If the configured burst value
is too low, achieved bandwidth limit will be lower than expected.
If the configured burst value is too high, too few packets could
be limited and achieved bandwidth limit would be higher than expected.

So we should recommend a correct example in case that user ignores
the note.

Change-Id: Iefea3ce699c39e217e89dec93a8fe8ea1a90ac82