Build metadata_agent.ini reproducibly
Currently, when metadata_agent.ini is built, the default value for the
directive metadata_workers is the build host's number of CPU. This is
wrong because metadata_agent.ini cannot be built reproducibly, which is
a bug in many distributions.

See for Debian: https://wiki.debian.org/ReproducibleBuilds/About

This patch therefore uses sample_default oslo.config directive to
hardcode a value put in generated configuration file that would not
depend on build environment.

Change-Id: I7292d09b96f90d0477dd4b59766854a733e1da38
Closes-Bug: #1736114