Add/rm centralized fips for DVR+HA from standby node
For DVR+HA:
When adding/removing a centralized fip, just add/remove it to/from
keepalived instance and do not check if it is a master router.

When adding a centralized fip, it is still needed to invoke method
of supperclass to configure nat rules on master node.

When removing a centralized fip, it is still needed to invoke method
of supperclass to clean contrack on master node.

Change-Id: I202db59acb7fb0b7924748cf7def2879a748e56a
Closes-Bug: #1736068
(cherry picked from commit 2200b1c3697a478af9f25a20ce09564bc2c9a968)