Use hexadecimal when adding DSCP rules
The iptables_manager can use iptables-save/restore for
manipulating rules, therefore we need to always use that
format when constructing them.

Found through inspection that the DSCP marking rules were
being added as integers, but iptables-save shows them in
hexadecimal:

  --> iptables -t mangle -A INPUT -j DSCP --set-dscp 12
  --> iptables-save | grep DSCP
  -A INPUT -j DSCP --set-dscp 0x0c

Adjust the rule string accordingly so that they are matched
correctly when the iptables_manager updates the rules.

Trivialfix

Change-Id: I2b3c37810d3dda0bc3ddeab0a84fee5d8cbe306c