Fix URL in SR-IOV internals doc
The last link in the 'Further Reading' section of the SR-IOV internals
document is not rendered properly. It gets a '/' at the end, that
makes the link useless. It seems to be a problem with Sphinx. To
work around it, the bullets are removed from the list of recommended
documents for further reading

Closes-Bug: #1737833

Change-Id: I2116e82d60bc361583eab8e3809ec5f9eab6d11c