Introduce rfe-confirmed and rfe-triaged tags
Right now we use Launchpad State "Confirmed" and "Triaged" for
the query to generate the agenda of neutron-drivers meetings.
Unfortunately it's often interfered by the gerrit integration.
E.g. A submission of Closed-Bug patch changes the state to "In Progess".

This patch avoids the issue by tweaking the procedure to use a set of
tags instead.

An altenative would be to teach the gerrit integration about our usage.
Or, move to some other systems like storyboard.
But they would require more efforts than this change.

Change-Id: I9becc8052b80a284a4cb336e2da115d28d501720