FIP: update_floatingip_status() to return the updated object
Prior to this patch the update_floatingip_status() method didn't have a
return value. This patch changes that by returning the updated version
of the floating ip upon update success, just like any other update()
method.

This is important in certain situations because,
update_floatingip_status() bumps the ``revision_number`` of the FIP.
Drivers like networking-ovn (and potentially others) does track those
revision numbers updates as a way to ensure that the SDN database is up
to date with the Neutron database.

Change-Id: Ic261e91761ff8715c227f5213870d2db014355c8