Speed up trunk MTU enforcement check
While things may have become generally slower for reasons still
to be determined at the time this patch was developed, we noticed
that enforcing MTU rules between subports and trunk parent was
done by performing one DB lookup per subport. This patch optmizes
the performance by piggybacking on the logic that is used to
retrieve the subport underlying network's segmentation type when
the segmentation type must be inherited.

Closes-bug: #1741954

Change-Id: Ib18427ef7a3509088eaf8fa0cfed3fb659b6baea