remove unused common rpc apis
Today neutron.common.rpc has a handful of public constants and functions
that are not used outside of neutron. This patch removes those not
externally used and enhances the rpc.init function to accept extra
RPC exception modules (though it's not used today) to still support
the notion of "extra mods" for RPC.

This patch also paves the way for easier rehoming the module to
neutron-lib.

Change-Id: I51607216a39980c17807dcccace3ab66b549607b