ovs-fw: Don't modify passed rules to update
When rules are passed to update() method of firewall, it translates
protocol in case it was passed as a string to integer. This can cause a
side-effect in testing when same set of rules is passed multiple times.
Securiry group rules always have a string in protocol field.

Change-Id: I0d95c6c41ad3c4b52986307449d90f13bbd412ac
(cherry picked from commit 37aba76275fd9a867ac15aa5c503b602b469ff97)