Fixing the filter in get MTU by network list query.
Based on the observation that a call to sync_routers can be slow
(10-15 seconds for one virtual router), and that profiling data show
that a significant amount of time is spent in function
_get_mtus_by_network_list, this change fixes incorrect the database
query for getting information about networks.

What changed:
- before: SQL query spent time O(#networks in database)
          (select all exists networks from database)
- after : SQL query spent time O(#network_ids passed to the function)
          (select only needed networks)

A basic test shows a drastic improvements, from 2 second to around
~0.2 second, in the processing of a _get_mtus_by_network_list call
with 300 networks in database.

Change-Id: Id50d1aaddd0d2a1110502de9c5b974063daa296a
Closes-Bug: #1748888