[Fullstack] Limit number of Neutron's api workers
Default number of api workers in Neutron is set to be equal to
number of CPU cores on host. That is fine on production environment
but on fullstack tests, where each test spawns own neutron-server
process it might cause host overload.

This patch limits number of api_workers to 2 which should be enough
for single test case and should make significantly lower load on host.

Change-Id: I1e970e35883d5240f0bd30eaea50313d93900580
Closes-Bug: #1750337