[Fullstack] Limit number of test workers to 4
By default there is as many test workers running fullstack
tests as cores on host.
On gate tests it is 8 cores so 8 test workers.
Each such workers spawns own neutron-server and agents.
That cause quite big load on such test node. From my tests
it looks that it can be up to 25.
By lowering number of test workers to 4 I observed that
load is at max about 10.
Overall time of tests is longer for about 10-15 minutes
(about 45-50 minutes with 8 workers and about 60-65 minutes
with 4 workers).

Change-Id: I58cb2ade598e8dc2377e6187483cb39881a9dc3c