Update upper-constraints consumption
In order to deal with neutron going into the upper-constraints file so
that neutron plugins can consume it normally, we need to align the
constraints invocation in neutron's tox.ini. This means putting the -C
argument into the dependencies list rather than the pip install line so
that it doesn't get appended to tox installing neutron itself.

Needed-By: https://review.openstack.org/550475
Change-Id: I014177c8afcadd3e6acb20bd38bb9d9fcc778ed6