Add search capability to the docs
This follows the nova docs and adds an in-tree neutron-specific
docs search and an openstack-wide search.

The former is necessary for finding things that aren't obviously
linked from top level indexes, like documentation about the
metadata service proxy (which is what I was originally trying
to find without the help of a search box).

Change-Id: I86dde1936d0c7d445f8a6db182f5f41dfb179752