DVR: Restarting l3 agent loses centralized fip ip on qg-interface
When l3 agent is restarted on a dvr_snat node that is configured
for L3_HA and has a centralized FloatingIP configured to the
qg-interface in the snat_namespace, that FloatingIP is not
re-configured to the qg-interface when agent starts.

The reason being, the cidr is not being retrieved from the
keepalived instance and only retrieved from the
centralized_fip_cidr_set.

If 'L3_HA' is configured we need to retrieve it from the keepalived
instance.

This patch fixes the problem by retrieving the cidrs from the
keepalived instance for the qg-interface.

Change-Id: I848a20d06e2d344503a4cb1776dbe2617d91bc41
Closes-Bug: #1740450
(cherry picked from commit 64028a389ff904f15e471b44bd5b3979c5db2cd2)