Fix DHCP isolated subnets with routed networks
After the merge of [1], the DHCP agent will fail to configure a routed
network when attempting to generate host routes for a non-local subnet
that is also isolated from the point of view of the metadata service.
This patch adds a check to make sure that the host route is not added
for non-local subnets.

[1] https://review.openstack.org/#/c/468744

Change-Id: Ia03f538a7d2d10d600d9359da5b3a74532709d1f
Closes-Bug: #1758952
(cherry picked from commit bb5138cff4e6c383d4d9e902423072a1493832b9)