[L3][QoS] Cover mixed dvr_snat and compute node dvr router
The dvr_snat node can also run the nova-compute which will
make the fip_qos l3 agent extension failed to install the
qos tc rules for the dvr router floating IPs.

To cover such mixed deployment, the tc rules for floating IP
qos will be processed again in qrouter-namespace of the
dvr_snat node, and the device is rfp-device.

Change-Id: I9eca25201e0647d4da77b325fba03cca1db5bb4d
Closes-Bug: #1758316