Change ha_state property to always return a value
Right now, ha_state could return any value that is in
the state file, or even '' if the file is empty.  Instead,
return 'unknown' if it's empty.

We also need to update the translation map in the HA code
to deal with this new value to avoid a KeyError.

Related-bug: #1755243

Change-Id: I94a39e574cf4ff5facb76df352c14cbaba793e98