unit test: unbreak test cases for callbacks.subscribe
patch set of [1] added callback priority so that a new argument, priority,
was added to subscribe method. Thus its signature was changed.
It caused to break some unit tests.[2] Those test cases check arguments passed
to subscribe method.

The fix to those test cases needs to be compatible with older version of
neutron-lib.
So add a helper method, get_subscribe_args, to produce argument to subscribe
by checking if neutron-lib supports callback priority or not. and apply it
to broken test cases.

[1] https://review.openstack.org/#/c/541766/
[2] http://logs.openstack.org/periodic/git.openstack.org/openstack/neutron/master/openstack-tox-py35-with-neutron-lib-master/66d328a/testr_results.html.gz
  test_capabilities.CapabilitiesTest.test_register
  test_ovs_capabilities.CapabilitiesTest.test_register
  test_backend.ServerSideRpcBackendTest.test___init__

Change-Id: I94066ff8a8dae5c5637478aa52883b3b451f2857