Allow sub-resources to have standard attributes
Prior to this change a DB model with standard
attributes could declare that it was mapping to
an API resource, but not declare a mapping to a
sub-resources (see bug 1763347).

This change allows DB models with standard attributes
to advertise that they map API *sub*-resources, and
modifies the code that extends DB resources to support
these specific declarations.

Closes-Bug: 1763347
Needed-By: I77ce46c0f33e2a366076d51ce6586fb3008dc6b1

Change-Id: I7630aab5e4f38d0fba862adc2426d4a7ca04a679