Do not remove conntrack jump rules if no zone
In corner cases, the firewall code could try and remove
non-existent conntrack zone jump rules if a zone has never
been allocated.  This could happen on an agent restart
when there are no longer ports in the zone on the
compute node.  Skip the removal since it will just generate
an iptables warning complaining the existing rule does
not exist.

Change-Id: Ie32733b4a06b6d75cf1eb78915a510a4bb78f619
Closes-bug: #1765208