remove common db alaises to db utils
A handful of the functions implemented in neutron.db._utils are exposed
in common_db_mixin as aliases. These are unneeded as they're already
defined elsewhere.

This patch removes the following aliases in common_db_mixin:
- safe_creation
- model_query_scope
- model_query
- resource_fields

It also renames the UT for common db mixin to
test__utils to reflect the fact that its testing functions
from the _utils db module.

NeutronLibImpact

Change-Id: I53cd533f107d950f163ad5b19b47097546a4691d