Change metadata agent to log message after failure
If the metadata agent fails to send its report_state
to the server, it logs an exception:

   Failed reporting state!: MessagingTimeout: Timed out...

If it then tries a second time (on 30 second intervals)
it just goes on happily.  It would be nice if it logged
that it had success on the subsequent attempt so someone
looking at the logs know it recovered.

Change-Id: I7a826811a889c618a39a8a91c8157e1805a327eb